requirements.txt
matplotlib==3.7.1
numpy==1.23.5
pandas==2.0.1
scikit_learn==1.2.2
tensorflow_cpu==2.12.0
